Fixed Jaccard Similarity bullshit

+++ src/main/java/snippets/scoring/strategy/JaccardSimilarityStrategy.java
@@ -36,8 +36,8 @@ public class JaccardSimilarityStrategy implements Similarity {
     Double count1, count2, minCount = 0.0, maxCount = 0.0;
     Set<String> set = Sets.union(qv.keySet(), dv.keySet());
     for (String key : set) {
-      count1 = qv.containsKey(key) ? qv.get(key) : 1.0;
-      count2 = dv.containsKey(key) ? dv.get(key) : 1.0;
+      count1 = qv.containsKey(key) ? qv.get(key).doubleValue() : 1.0;
+      count2 = dv.containsKey(key) ? dv.get(key).doubleValue() : 1.0;
       minCount += (count1 < count2 ? count1 : count2);
       maxCount += (count1 > count2 ? count1 : count2);
     }
